phishing-hunter/requirements.txt
flask
pillow
pytesseract
opencv-python
scikit-learn
numpy
tensorflow
